refactor: drop coordinator resource package
